// src/lib/generated/deployments.ts
// AUTO-GENERATED by scripts/sync-abi.mjs — do not edit.
export const localDeployment = {
  "chainId": "0x0000000000000000000000000000000000007A69",
  "chainIdNum": 31337,
  "deployer": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
  "factory": "0x5FC8d32690cc91D4c39d9d3abcBD16989F875707",
  "network": "anvil",
  "oracle": "0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512",
  "pDGEN": {
    "assetSymbols": [
      "DOGE",
      "SHIB",
      "PEPE",
      "WIF"
    ],
    "assets": [
      "0x5081a39b8A5f0E35a8D959395a630b68B74Dd30f",
      "0x4C4a2f8c81640e47606d3fd77B353E87Ba015584",
      "0xDC11f7E700A4c898AE5CAddB1082cFfa76512aDD",
      "0x0355B7B8cb128fA5692729Ab3AAa199C1753f726"
    ],
    "token": "0xB2d7EdA927332Ec84eCA5B9a6b23B4F45335E8e1",
    "vault": "0x3dE2Da43d4c1B137E385F36b400507c1A24401f8"
  },
  "pMETL": {
    "assetSymbols": [
      "SLV",
      "PPLT",
      "COPX"
    ],
    "assets": [
      "0x809d550fca64d94Bd9F66E60752A544199cfAC3D",
      "0xb7278A61aa25c888815aFC32Ad3cC52fF24fE575",
      "0x7969c5eD335650692Bc04293B07F5BF2e7A673C0"
    ],
    "token": "0x8C1f1Ab0b323c4B4FCb9F85CE8beCbCD6F673d43",
    "vault": "0xb0279Db6a2F1E01fbC8483FCCef0Be2bC6299cC3"
  },
  "pSEMI": {
    "assetSymbols": [
      "NVDA",
      "AMD",
      "AVGO",
      "TSM",
      "ASML",
      "MU",
      "QCOM",
      "INTC"
    ],
    "assets": [
      "0x0165878A594ca255338adfa4d48449f69242Eb8F",
      "0x610178dA211FEF7D417bC0e6FeD39F05609AD788",
      "0x9A676e781A523b5d0C0e43731313A708CB607508",
      "0x68B1D87F95878fE05B998F19b66F4baba5De1aed",
      "0x4ed7c70F96B99c776995fB64377f0d4aB3B0e1C1",
      "0x7a2088a1bFc9d81c55368AE168C2C02570cB814F",
      "0xE6E340D132b5f46d1e472DebcD681B2aBc16e57E",
      "0xa82fF9aFd8f496c3d6ac40E2a0F282E47488CFc9"
    ],
    "token": "0x63f84713F52422Af2F8E18b56703b0f80CCcCBcE",
    "vault": "0x856e4424f806D16E8CBC702B3c0F2ede5468eae5"
  },
  "swapExecutor": "0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9",
  "usdg": "0x5FbDB2315678afecb367f032d93F642f64180aa3"
} as const;